% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxBScut.R
\name{oxBScut}
\alias{oxBScut}
\title{oxBScut}
\usage{
oxBScut(matrix_5hmc, poobha_5hmc)
}
\arguments{
\item{matrix_5hmc}{is a matrix or dataframe of 5hmC beta values.
poobha_5hmc is a matrix or dataframe of p-values with out-of-band array 
hybridization that is corresponding to the 5hmC beta matrix.}
}
\value{
Recommended 5hmC cut-off, low-quality fraction, a figure on 5hmC distribution
between high-quality probe and low-quality probe, a figure on low-quality
CpG genomic location enrichment, a figure on coefficient of variation and 
median 5hmC with the recommended 5hmC cut-off marked.
}
\description{
The function makes a recommendation for a 5hmC cut-off to eliminate unstable
5hmC probes on the Illumina HumanMethylation BeadChip and evaluate the 
information loss with the low-quality probes.
}
