% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxBSlabel.R
\name{oxBSlabel}
\alias{oxBSlabel}
\title{oxBSlabel}
\usage{
oxBSlabel(rgSet)
}
\arguments{
\item{rgset}{'RGChannelSet' including pheno matrix with columns "Sample_ID" that is
unique to each sample and “Subject" that is unique to each paired subject.}
}
\value{
A pheno matrix with columns on predicted BS and oxBS label.
}
\description{
The function predicts the BS or oxBS label for paried bisulfite and oxidative
bisulfite treated samples using the internal normalization control probes on
the Illumina HumanMethylation BeadChip.
}
